+++ .storybook/preview.ts
@@ -21,6 +21,7 @@ import '../packages/ui/dist/css/components/utility-sans.css';
 import '../packages/ui/dist/css/components/utility-serif.css';
 import '../packages/ui/dist/css/components/image.css';
 import '../packages/ui/dist/css/components/prose.css';
+import '../packages/ui/dist/css/components/list.css';
 
 export const parameters = {
   controls: {