+++ packages/lint/src/index.ts
@@ -1,6 +1,7 @@
 import buttonRules from './rules/button-rules.js';
 import containerRules from './rules/container-rules.js';
 import imageRules from './rules/image-rules.js';
+import listRules from './rules/list-rules.js';
 import linkRules from './rules/link-rules.js';
 import proseRules from './rules/prose-rules.js';
 import textRules from './rules/text-rules.js';
@@ -9,6 +10,7 @@ const rules = {
   ...buttonRules.rules,
   ...containerRules.rules,
   ...imageRules.rules,
+  ...listRules.rules,
   ...linkRules.rules,
   ...proseRules.rules,
   ...textRules.rules,
@@ -20,6 +22,7 @@ const configs = {
       ...buttonRules.configs.recommended.rules,
       ...containerRules.configs.recommended.rules,
       ...imageRules.configs.recommended.rules,
+      ...listRules.configs.recommended.rules,
       ...linkRules.configs.recommended.rules,
       ...proseRules.configs.recommended.rules,
       ...textRules.configs.recommended.rules,
@@ -41,6 +44,8 @@ export type {
   ContainerTagAnalysis,
   ImageClassInfo,
   ImageTagAnalysis,
+  ListClassInfo,
+  ListTagAnalysis,
   LinkClassInfo,
   LinkTagAnalysis,
   ProseClassInfo,