+++ sandbox/vue-library/src/index.ts
@@ -1,6 +1,7 @@
 import CedarButtonFixture from './components/CedarButtonFixture.vue';
 import CedarContainerFixture from './components/CedarContainerFixture.vue';
 import CedarImageFixture from './components/CedarImageFixture.vue';
+import CedarListFixture from './components/CedarListFixture.vue';
 import CedarLinkFixture from './components/CedarLinkFixture.vue';
 import CedarProseFixture from './components/CedarProseFixture.vue';
 import CedarTextFixture from './components/CedarTextFixture.vue';
@@ -10,6 +11,7 @@ export {
   CedarButtonFixture,
   CedarContainerFixture,
   CedarImageFixture,
+  CedarListFixture,
   CedarLinkFixture,
   CedarProseFixture,
   CedarTextFixture,